LPS-164755 Refactor for simplicity
